Updated VS project file's Release parameters
Before the initial upload to the repository, the project file was recreated.  The new file was only tested in Debug mode so some settings required for Release mode were overlooked.  The changes include changing the project configuration type from 'Application' to 'StaticLibrary', as well as adding the Qt SVG module.