WIP: Migrating Layers core classes away from Qt
- A majority of Qt constructs have been removed from Layers core
classes
- Theme loading/saving is now handled interally with JSON functionality
included with Layers.
- Theme UUIDs are now generated by Layers
- Theme IDs can now be made up of a composite of the theme's name and
the known publisher of the theme
- For local custom themes, the username of the user that is currently
logged in on the system is recorded as the publisher. In the future,
when users can publish themes to a Layers online network of sort, the
publisher will be changed to their online username.
- The Dark and Light default themes are no longer handled through Qt
resources.
- Implemented LAttribute::as_if which returns a pointer to the
attribute's value of template type. If the value is not of the template
type, then nullptr is returned.