Split Layers into Layers and QLayers
- Layers is now an independent theme framework that can be integrated
with other application frameworks. Users of the Layers library need to
implement a subclass of LThemeable and use it to integrate Layers with
widget types to make them capable of applying and utilizing Layers
themes.
- QLayers is a Layers-SDK/translation-layer (still trying to figure out
the right term for it) that provides functionality for integrating
Layers with Qt applications. It includes all of the previously
implemented Layers widgets. QLayers has been removed from this repo
since an independent repo will be created for it.